requirements.txt
streamlit>=1.32.0
openai>=1.12.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
openpyxl>=3.1.0
